feat!: drop webgl2 fallback
Removes the WebGL2 fallback renderer. The embedding view now requires WebGPU.

BREAKING CHANGE: browsers without WebGPU are no longer supported — the status bar reports it instead of falling back.

- Delete webgl2_renderer/ and the fallback path in EmbeddingViewImpl.svelte.
- Delete renderer_interface.ts; its types move into webgpu_renderer/renderer.ts and EmbeddingRendererWebGPU becomes EmbeddingRenderer. Not public API.
- maxDensityModeCategories() returns 32 unconditionally; the 4-category cap was a WebGL2 limit.
- Fold isWebGPUAvailable into requestWebGPUDevice.
- Remove a dead Intl.Segmenter availability check in viewer/src/embedding/scorer.ts.
