+++ packages/component/src/lib/webgpu_renderer/renderer.ts
@@ -10,8 +10,7 @@ import {
   type Matrix3,
   type Vector4,
 } from "../matrix.js";
-import type { DensityMap, EmbeddingRenderer, EmbeddingRendererProps, RenderMode } from "../renderer_interface.js";
-import type { ViewportState } from "../utils.js";
+import type { Point, ViewportState } from "../utils.js";
 import { Viewport } from "../viewport_utils.js";
 import { makeModuleUniforms, type ModuleUniforms } from "./uniforms.js";
 import { gpuBuffer, gpuBufferData, gpuTexture } from "./utils.js";
@@ -27,7 +26,51 @@ import { kdeConfig } from "./kde_config.js";
 
 import programCode from "./program.wgsl?raw";
 
-export class EmbeddingRendererWebGPU implements EmbeddingRenderer {
+export type RenderMode = "points" | "density";
+
+export interface EmbeddingRendererProps {
+  mode: RenderMode;
+  colorScheme: "light" | "dark";
+
+  x: Float32Array<ArrayBuffer>;
+  y: Float32Array<ArrayBuffer>;
+  category: Uint8Array<ArrayBuffer> | null;
+
+  categoryCount: number;
+  categoryColors: string[] | null;
+
+  viewportX: number;
+  viewportY: number;
+  viewportScale: number;
+
+  pointSize: number;
+  pointAlpha: number;
+  pointsAlpha: number;
+
+  densityScaler: number;
+  densityBandwidth: number;
+  densityQuantizationStep: number;
+  densityAlpha: number;
+  contoursAlpha: number;
+
+  gamma: number;
+  width: number;
+  height: number;
+
+  /** Approximate maximum points to render. null/Infinity = no limit. Default: 4,000,000 */
+  downsampleMaxPoints: number | null;
+  /** Density weight for downsampling (0-10). Default: 5 */
+  downsampleDensityWeight: number;
+}
+
+export interface DensityMap {
+  data: Float32Array;
+  width: number;
+  height: number;
+  coordinateAtPixel: (x: number, y: number) => Point;
+}
+
+export class EmbeddingRenderer {
   readonly props: EmbeddingRendererProps;
 
   private viewport: Viewport;
@@ -110,6 +153,7 @@ export class EmbeddingRendererWebGPU implements EmbeddingRenderer {
     );
   }
 
+  /** Set renderer props. Returns true if a render is needed. */
   setProps(newProps: Partial<EmbeddingRendererProps>): boolean {
     let needsRender = false;
     let key: keyof EmbeddingRendererProps;
@@ -146,14 +190,17 @@ export class EmbeddingRendererWebGPU implements EmbeddingRenderer {
     return needsRender;
   }
 
+  /** Render */
   render(): void {
     this.renderer.value(this.props, this.context.getCurrentTexture().createView());
   }
 
+  /** Destroy the renderer and free any resource */
   destroy(): void {
     this.df.destroy();
   }
 
+  /** Produce a density map */
   async densityMap(width: number, height: number, radius: number, viewportState: ViewportState): Promise<DensityMap> {
     let subgraph = this.df.subgraph();
     let { x, y, scale: s } = viewportState;
